Fix issue with past links in github pages deploy
Problem
--------

Because when we deploy to gh pages, links are no longer based off of `/`
but are nested under our repo name, in this case `/c5-zine` like
`/c5-zine/issue4`.  This works fine for the first link, but once you're
on a past issue (`/c5-zine/issue4`) when you click the link for
*another* past issue, you get `/c5-zine/issue4/issue3`.

This case is specific to gh-pages deployment.  We cannot know the
deployment strategy when the app runs because it's static.  If we build
links that will work for gh-pages, then local development with the
middleman server don't work.

See Issue #8

Solution
---------

We can use a little bit of javascript that looks at the current path
when you're browsing the page.  If that path has more than 1 level, we
use that path to generate the new past link path by replacing the last
part of the path with the new href.

This should work unless we start adding some deeper paths than the
top-level.

Changes
-------

* Move javascript from index.html.erb into the `site.js`
* Add a listener that is attached to all `past-issue__link`s which
  builds the past issue link based on the current path and the desired
  href and intercepts the click to take the user to the right place